+++ test/toStructure.test.js
@@ -8,7 +8,7 @@ var toStructure = require('../lib/toStructure'),
 (function () {
 	var source = '<html><head><title>Something &amp; Else</title></head><body class="test"><!-- A comment  --></body></html>';
 	
-	toStructure(source, {}, function(result){
+	toStructure(source, {}, function(err, result){
 		module.exports['structure can include tags with children'] = function() {
 			assert.equal(result.length, 1);
 			assert.equal(result[0].children.length, 2);
@@ -42,33 +42,23 @@ var toStructure = require('../lib/toStructure'),
 	
 	source = '<html><head><title>Something</head><body></body></html>';
 	
-	toStructure(source, { callbackErrors:true }, function(result){
+	toStructure(source, { callbackErrors:true }, function(err, result){
 		module.exports['structure should produce error for unclosed title tag'] = function() {
-			assert.equal(result.length, 1);
-			var element = result[0];
-			assert.equal(element.type, 'error');
-			assert.equal(element.element, 'title');
-			assert.equal(element.line, 1);
-			assert.equal(element.column, 45);
+            assert.ok(err);
 		};
 	});
 	
 	source = '<html><head><title var>Something</title></head></html>';
 	
-	toStructure(source, { callbackErrors:true }, function(result){
+	toStructure(source, { callbackErrors:true }, function(err, result){
 		module.exports['structure should produce error for attributes without values'] = function() {
-			assert.equal(result.length, 1);
-			var element = result[0];
-			assert.equal(element.type, 'error');
-			assert.equal(element.element, 'head');
-			assert.equal(element.line, 1);
-			assert.equal(element.column, 28);
+		    assert.ok(err);
 		};
 	});
 	
 	source = '<html><head attribute="%{ Something }"><title><%= Something %></title></head></html>';
 	
-	toStructure(source, {}, function(result){
+	toStructure(source, {}, function(err, result){
 		module.exports['structure should produce ejs tag within title tag'] = function() {
 			var element = result[0].children[0].children[0];
 			assert.equal(element.type, 'tag');
@@ -81,7 +71,7 @@ var toStructure = require('../lib/toStructure'),
 	
 	source = '<html><head attribute="%{ \'Something\' }"><title><%= "Something" %></title></head></html>';
 	
-	toStructure(source, {}, function(result){
+	toStructure(source, {}, function(err, result){
 		module.exports['structure can tolerate double-quotes in ejs tags'] = function() {
 			var element = result[0].children[0].children[0];
 			assert.equal(element.type, 'tag');